+++ Pandora/Paradigm/Structure/Ability/Monotonic.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Structure.Ability.Monotonic where
 
 class Monotonic e a where
-	iterate :: (a -> r -> r) -> r -> e -> r
+	bypass :: (a -> r -> r) -> r -> e -> r
 
 instance Monotonic a a where
-	iterate f r x = f x r
+	bypass f r x = f x r